+++ src/base/b-dummy/b-dummy.ts
@@ -10,8 +10,8 @@ import { inViewFactory } from 'core/dom/in-view';
 import { ImageLoader, imageLoaderFactory } from 'core/dom/image';
 import { ResizeWatcher } from 'core/dom/resize-observer';
 
+import iLockPageScroll from 'traits/i-lock-page-scroll/i-lock-page-scroll';
 import iData, { component } from 'super/i-data/i-data';
-import iInputText from 'super/i-input-text/i-input-text';
 import { Directives, Modules } from 'base/b-dummy/interface';
 
 const
@@ -28,7 +28,17 @@ export * from 'base/b-dummy/interface';
 	}
 })
 
-export default class bDummy extends iInputText {
+export default class bDummy extends iData implements iLockPageScroll {
+	/** @see [[iLockPageScroll.lock]] */
+	lock(): Promise<void> {
+		return iLockPageScroll.lock(this);
+	}
+
+	/** @see [[iLockPageScroll.unlock]] */
+	unlock(): Promise<void> {
+		return iLockPageScroll.unlock(this);
+	}
+
 	/**
 	 * Links to directives
 	 */