+++ src/core/dom/in-view/test/index.js
@@ -228,7 +228,9 @@ module.exports = async (page, params) => {
 				});
 
 				await delay(300);
+
 				expect(await page.evaluate(() => globalThis.tmp)).toBeUndefined();
+				expect(await getInView(strategy).evaluate((ctx) => ctx.adaptee.elements.size)).toBe(0);
 			});
 
 			it('suspending with `callback`: doesn\'t fire the callback', async () => {