+++ tests/cases.js
@@ -19,5 +19,6 @@ module.exports = [
 	'--test-entry base/b-list/test',
 	'--test-entry super/i-block/test',
 	'--test-entry core/dom/image/test',
-	'--test-entry core/dom/in-view/test'
+	'--test-entry core/dom/in-view/test',
+	'--test-entry traits/i-lock-page-scroll/test'
 ];