+++ src/base/b-router/CHANGELOG.md
@@ -23,6 +23,12 @@ Changelog
 
 * Improved documentation
 
+## v3.0.0-rc.110 (2020-12-16)
+
+#### :rocket: New Feature
+
+* Added `interceptLinks`
+
 ## v3.0.0-rc.36 (2020-07-13)
 
 #### :bug: Bug Fix