+++ src/base/b-sidebar/CHANGELOG.md
@@ -9,6 +9,18 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.0.0-rc.110 (2020-12-16)
+
+#### :rocket: New Feature
+
+* Added `forceInnerRender`
+
+## v3.0.0-rc.100 (2020-11-17)
+
+#### :house: Internal
+
+* Rendering optimization
+
 ## v3.0.0-rc.98 (2020-11-13)
 
 #### :boom: Breaking Change