+++ src/core/component/render-function/CHANGELOG.md
@@ -9,6 +9,18 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.0.0-rc.105 (2020-12-09)
+
+#### :bug: Bug Fix
+
+* Fixed a bug with redundant `v-for` invokes
+
+## v3.0.0-rc.99 (2020-11-17)
+
+#### :bug: Bug Fix
+
+* [Fixed dynamic creation of flyweight components](https://github.com/V4Fire/Client/issues/434)
+
 ## v3.0.0-rc.96 (2020-11-10)
 
 #### :rocket: New Feature