+++ src/super/i-block/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.0.0-rc.110 (2020-12-16)
+
+#### :rocket: New Feature
+
+* Added API to load the dynamic dependencies
+
 ## v3.0.0-rc.92 (2020-11-03)
 
 #### :nail_care: Polish