+++ src/super/i-block/modules/module-loader/CHANGELOG.md
@@ -9,8 +9,8 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
-## v3.0.0-rc.37 (2020-07-20)
+## v3.0.0-rc.110 (2020-12-16)
 
-#### :house: Internal
+#### :rocket: New Feature
 
-* Fixed ESLint warnings
+* Initial release