+++ src/super/i-block/modules/state/CHANGELOG.md
@@ -23,6 +23,12 @@ Changelog
 
 * Improved documentation
 
+## v3.0.0-rc.110 (2020-12-16)
+
+#### :boom: Breaking Change
+
+* Now `initFromStorage` returns `CanPromise`
+
 ## v3.0.0-rc.45 (2020-07-30)
 
 #### :bug: Bug Fix